Resolve "Implement Geometry::clone"

Closes #175

See merge request aerys/smartshape-engine!173

(cherry picked from commit 648b494d4f7c0e89bca33773a1b6fcf0e4bae06c)

dce0f61a Fix the debug build of the serializer plugin (a PR should be issued in the crunch library).
a0dc13cd Implement Geometry::clone.
2cc1a35d Fix indentation.
4ec2e2b4 Fix a typo and make more explicit comment.

+++ plugin/serializer/lib/crnlib/inc/crn_decomp.h
@@ -2530,7 +2530,7 @@ namespace crnd
          return NULL;
       }
 
-      CRND_ASSERT(((uint32)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
+      CRND_ASSERT(((uint64)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
 
       return p_new;
    }
@@ -2555,7 +2555,7 @@ namespace crnd
       if (pActual_size)
          *pActual_size = actual_size;
 
-      CRND_ASSERT(((uint32)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
+      CRND_ASSERT(((uint64)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
 
       return p_new;
    }
